feat: create themes for chat

+++ src/shared/lib/helpers/localStorageHelper/localStorageHelper.ts
@@ -1,5 +1,8 @@
+const isClient = typeof window !== 'undefined';
+
 export const localStorageHelper = <T extends string | object>(key: string) => {
 	const storageGetItem = (key: string) => {
+		if (!isClient || !localStorage) return {};
 		const raw = localStorage.getItem(key);
 		if (!raw) return {};
 		try {
@@ -9,12 +12,15 @@ export const localStorageHelper = <T extends string | object>(key: string) => {
 		}
 	};
 	const storageSetItem = (value: Partial<T>) => {
+		if (!isClient || !localStorage) return;
 		localStorage.setItem(key, JSON.stringify(value));
 	};
 	const storageRemoveItem = () => {
+		if (!isClient || !localStorage) return;
 		localStorage.removeItem(key);
 	};
 	const storageClear = () => {
+		if (!isClient || !localStorage) return;
 		localStorage.clear();
 	};
